add NotHandle sentinel error value
A maybe HTTP handlers is an HTTP handler that may or may not handle
a request, depending on whether it matches the necessary criteria.

This new sentinel error value will be used to by such maybe HTTP
handlers to report that the request was explicitly not handled.

Any other error, such as the nil error or another non-nil error,
implies the maybe HTTP handler has handled the given request.